Put the "Month" column inside a dict in the replace() method
Everytime I  was replacing the month number with the month name in df_months, I've been getting a warning from Pandas about referencing the "Month" column with square braces. Pandas recommended putting this column as a key in a dict and the value as the "month_map" variable, which contains the values to be replaced in the "Month" column. This little change must prevent any future error that Pandas might return in future updates