Change the way to reorder the categories in the plots
I found a simpler way to get the handles and labels to reorder them. Matplotlib has a function to get them once and return two lists, which I put in two variables: one for the handles and another for the labels. Then, I needed just to reorder these lists using indexing and pass them into the legend() method. With this, I could reorder the categories in both two plots